requirements.txt
pandas>=2.0
openpyxl>=3.1
openai>=1.0
requests>=2.31
pyyaml>=6.0
aiohttp>=3.9
httpx>=0.27
